Switch int to size_t to indicate that the value will never be lower than zero

+++ leetcode/cppinclude/mysemaphore.h
@@ -5,16 +5,16 @@
 
 class Semaphore {
 private:
-  int avail;
+  size_t avail;
   std::mutex m;
   std::condition_variable cv;
 
 public:
   /** Default constructor. Default semaphore is a binary semaphore **/  
-  explicit Semaphore(int avail_ = 1) : avail(avail_) {}
+  explicit Semaphore(size_t avail_ = 1) : avail(avail_) {}
   void wait();
   void post();
-  int available() const;
+  size_t available() const;
 };
 
 #endif //MYSEMAPHORE_H
